requirements.txt
python-lirc==1.2.3
pyserial==3.2.1
py_irsend==1.0.0
